+++ app/src/main/java/com/example/android_interfaces_java/Interfaces/AirTravel.java
@@ -1,4 +1,4 @@
-package com.example.android_interfaces_java;
+package com.example.android_interfaces_java.Interfaces;
 
 public interface AirTravel  {
     public void fly();
