+++ app/src/main/java/com/example/android_interfaces_java/Interfaces/GroundTravel.java
@@ -1,4 +1,4 @@
-package com.example.android_interfaces_java;
+package com.example.android_interfaces_java.Interfaces;
 
 public interface GroundTravel {
     public void drive();
