+++ app/src/main/java/com/example/android_interfaces_java/Interfaces/WaterTravel.java
@@ -1,4 +1,4 @@
-package com.example.android_interfaces_java;
+package com.example.android_interfaces_java.Interfaces;
 
 public interface WaterTravel {
     public void sail();
