+++ app/src/main/java/com/example/android_interfaces_java/PontoonBoat.java
@@ -2,6 +2,9 @@ package com.example.android_interfaces_java;
 
 import android.util.Log;
 
+import com.example.android_interfaces_java.Interfaces.WaterTravel;
+import com.example.android_interfaces_java.Vehicle;
+
 public class PontoonBoat extends Vehicle implements WaterTravel {
     @Override
     void travel() {
